FunctionApps/dev/Code/translateToHex/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.21.3
azure-core==1.32.0
# python-dotenv==1.0.1
# requests==2.32.3